Add ReceiptMetadata with verifier_info field for checking for verifier parameter mismatch (#1764)
This PR adds a `ReceiptMetadata` struct and attaches it to the `Receipt`
type.

The first field added to the `ReceiptMetadata` is a `verifier_info:
Digest` value, which is constructed from a hash of the parameters for a
given a given verifier.
This value can be used to tell when a `Receipt` and a verifier have
different parameters (e.g. the control root may have changed).
It is intented additionally to serve as a routing key for the Ethereum
on-chain verifier contracts, ensuring we send a given receipt to the
right place.

Additionally, this PR:

* Adds an `Input` type to the `ReceiptClaim` to serve as a placeholder
for a future value type.
It is implemented as an uninhabited type, which means it cannot be
constructed.
This ensures that adding fields to this type is not a breaking change to
our APIs, and all code paths will deal with the digest only.
* Refactors the `risc0_groth16` public API to introduce a new
`VerifyingKey` type.
This type implements the `Digestible` trait and is serializable with
`serde`, which eases implementation of `ReceiptMetadata`.
Additionally, it works to hide the `arkworks` types such that we have
the option to swap arkworks (which is primarly a library for R&D of
proof systems) out for a different Groth16 verifier implementation.