Enable composition of rv32im programs and generalized recursion programs (#1841)
As a major step towards composition with recursion programs and circuits
implementing accelerators, this PR adds a `control_root` value to the
receipt claim encoding of assumptions, and enables proving and of
recursion programs that are not in the main set of rv32im transforms.

In the course of this PR, there are also a number of changes to make it
possible to pass alternate parameters to proving and verification. In
particular, `ProverOpts` now holds the set of allowed control roots and
can be configured to use an alternate set, and `VerifierContext` holds
the verifier parameters for each receipt type.

`SuccinctReceipt` has been made more generic, allowing it to represent
proof of any recursion program that follows the convention of placing a
control root in the first output slot, and a SHA-256 claim digest in the
second output slot. This claim digest can be over a generic claim, which
allows for claims representing the execution of e.g. accelerators which
are not `ReceiptClaim` and do not represent zkVM execution.

Verification routines on `Receipt` and `CompositeReceipt` are
simplified, partially taking advantage of recent changes to the rv32im
circuit to zero-out the post-state digest, and partially by dropping
`ExitCode::Paused(0)` as a default accepted exit code.

---------

Co-authored-by: Frank Laub <flaub@risc0.com>