Fix `send_recv` for cases where `U != u8`
The `send_recv` mechanism works by making a pair of syscalls. One to get
the size of the data and another to retrieve that amount of data. There
was a mistake in assuming that the first syscall returned the number of
elements to return. In fact, the first syscall returns the number of
bytes. The guest handler has been fixed to reflect this.

The calculation for buffer size in the subsequent syscall has been fixed
by removing the multiplication to calculate the size of the the host
buffer and by dividing by `sizeof::<U>` before returning to the guest.