recursion circuit: create an empty zkr.zip file when building docs (#1907)
This fixes docs.rs build errors that we've been seeing for
`risc0-circuit-recursion` and hopefully `risc0-zkvm`. I was able to test
this locally and verify that `risc0-circuit-recursion` has been fixed
but `risc0-zkvm` fetches the published `risc0-circuit-recursion` crate
so changes to not appear in my local testing environment.