Add c-kzg tesst as an example
Some unknown issue (likely with cargo, workspaces, and or build scripts)
is causing the c-kzg guest to not build in the test directory. Add c-kzg
as an example as a temporary measure until we can get to the bottom of
why this crate won't build when placed under the `risc0/zkvm/methods`
directory.