Examples: alter user on version incompatibility
This is a work-in-progress commit intended to get early feedback.
I've attempted multiple approaches to implement this but I realized that
the solution we need right now must not modify the server because if the
server is modified, then the client code will not work. This approach
runs `r0vm --version` as a command and captures its output to determine
the version of the server. This seems to be the only approach to get the
server version without modifying the server code and is intended to be
compatible with https://github.com/risc0/risc0/pull/2104.

In the very near future, I intend to
1)create a `risc0-zkvm::get_server_version` function for `feature = client`
2)refactor code in the build script to a separate examples `util` module
  and call that from every build script in the examples and do this in
  a way that the version check will be disabled when -F prove is present

```
       WARNING: The example code that you are trying to run is not compatible with your
       installation of RISC Zero and the cargo command you ran to build or run this example
       will be terminated. Run the command below to checkout the correct version of the
       example code and run your cargo command again.

                     git checkout origin/release-1.0

       Version of the current example code:    1.1.0-alpha.1 (pre-release)
       Version of your RISC Zero installation: 1.0.3
```