fix: enable `std` for `hex` (#2189)
This PR fixes building `cargo-risczero` without default features. This
is currently causing CI failures in
https://github.com/risc0/risc0-ethereum.

Using `?` to convert a `FromHexError` into an `anyhow:Error` requires
the `std` feature of the the `hex` crate. This happens in the following
line:

https://github.com/risc0/risc0/blob/b57db72db8c79a6f6bc52ad581189d167ba7eeef/risc0/cargo-risczero/src/commands/verify.rs#L73

This currently leads to compilation failures when building
`cargo-risczero` without default features, otherwise the correctly
configured `hex` will be pulled in via some other dependencies.

Since `std` is the only feature enabled with `default` in hex, removing
`default-features = false` seams preferable to explicitly adding the
`std` feature.