Back-port: Include powers of two (po2) up to 21 in the default verifier parameters (#2276) (#2281)
As part of our continual security analysis work, we've created a
calculated for the "bits of security" given all the system parameters
relevant to the soundness of the proof of knowledge property guaranteed
to the verifier. With this, we find that the lookup argument may not
achieve our targeted 97 bits of soundness under worse case assumptions
grounded in the Schwartz-Zippel lemma. Under these assumptions, up to 3
bits of security are given up by including segment sizes up to power of
two 24, relative to po2 21.

This PR implements the cautious approach of removing po2s above 21 from
the default verifier parameters. Developers that wish to use po2 22, 23,
or 24 may opt-in by using a non-default `VerifierContext`.

Additionally, this PR fixes a pair of off-by-one errors that had removed
the rv32im control IDs for po2 24. Technically, this constitutes a
breaking change in the `risc0-circuit-rv32im` crate since the public
type `ControlIds` is not a fixed array of length n+1.


https://github.com/risc0/risc0/pull/2276/files#diff-ce2293a7cee10116f619bedf007ecfedd524505b764cfe5383785b7bd17dd7c4

Note: It is expected that the `risc0-ethereum` CI job will fail right
now, because the control root is changed. It will pass again once we've
merged this and https://github.com/risc0/risc0-ethereum/pull/205 to
their respective `main`.

---------

Co-authored-by: Victor Graf <victor@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>