Support for deferred assumptions (#2288)
This change defers the processing of on-demand ZKR proofs so that they
occur later in the pipeline rather than inline during the execution
phase. The benefit of this approach is that proving services can
distribute the ZKR proof workload out to a worker pool.

A `CoprocessorCallback` can be configured on the `ExecutorEnv` which
allows users to receive a notification when a ZKR proof needs to be
produced. The ZKR proof can happen asynchronously and does not need to
block the executor. However, the assumption receipt needs to be produced
by the time `resolve` is called to resolve any outstanding assumptions.