Fix note about paging in RISC-V cycle table (#2193)
In the table of cycle counts per instruction, in the optimization guide,
there is a note that says "1 if [paged-in](#paging) 1094 to 5130
otherwise". In the case of load instructions, this is true. In the case
of store instructions, this is slightly inaccurate. It's instead based
on whether the memory is _dirty_ rather than if it is paged in.