Allow static mut references in specific places (#2398)
In recent nightly builds, there is a warning against making mutable
references to mutable statics (see e.g. here:
https://github.com/risc0/risc0/actions/runs/11148604622/job/30985587956?pr=2371).
We are doing so in certain `unsafe` blocks where we have confirmed that
our usage is acceptable, so this PR adds ` #[allow(static_mut_refs)]` to
those locations to suppress this warning.