Makes Bonsai Dependencies Optional under client feature flag (#2375)
The zkvm `client` feature flag includes the `bonsai-sdk`, which in turn
includes the `bonsai-sdk` dependancy tree.
Because the `bonsai-sdk` includes `reqwest` and its `rustls-tls` users
may find that they have version conflicts with various low level
packages such as `zeroize` and other cryptography related crates. This
is especially henious when working with other blockchain sdk crates such
as solana.

I propose the following change which is to make the bonsai-sdk only
active when a developer specifically asks for it. This change will
reduce compiled binary size, allow devs to unshackle themselves from
forked versions of risc0 such as
https://github.com/anagrambuild/risc0/tree/v1.0.1-bonsai-fix, and allow
devs implementing their own hosts to not have to bleed so much.

Its likely that some docs need to be updated and perhaps examples aswell
in order to fully integrate this change, but this PR is meant to see if
the Risc0 team will accept this type of change.

Note the `bonsai` flag will default to on so that current sdk
consumoooors will not be affected.