Expand RSA to 4096 bits in bigint2 (#2545)
Adjust modpow_65537 to work with 4096 bit values instead of 3072 bit
values. Smaller bitwidths are still supported (as long as they are
zero-padded to 4096).

This comes at a mild perf cost, and so in the long run we want more
optionality here, or perhaps a different approach (modmul-based). But
for now, this increases compatibility.

Based on https://github.com/risc0/zirgen/pull/88