Always have some default version if any version is installed
Due to having to deal with installs where there is no set active
version of a component (because there is no settings.toml) we have to
somehow pick a version from the installed versions.

This means that there should always be a default version if there is
some version installed. We don't modify the settings.toml file in the
path where we are finding toolchains.