Add some benchmarks to the datasheet for bigint2
This adds two benchmarks, one for execute and one for prove_segment.

The guest program runs k256 ECDSA verify in a loop.

The execute benchmark does about 1M cycles like the existing benchmark.

The prove_segment benchmark proves one full segment of this.

We use po2=20, which is the default