ZKVM-1028: update hello-world tutorial doc  (#2789)
As pointed out in the
[Discord](https://discord.com/channels/953703904086994974/1334542837387231242),
the hello world tutorial on the docs was not up to date. Specifically it
pointed the reader to use `cargo risczero new` as oppose to checking out
the correct `examples/hello-world` directory. It also had outdated code.

This PR updates the [hello world
tutorial](https://dev.risczero.com/api/zkvm/tutorials/hello-world) by:

* by guiding the reader to the right example code
* walking through the current hello world example code 
* running the example locally 

CC @pdg744