Only count a "push" if we are jumping to the start of a function
We have to consider some jumps as function calls because sometimes
function calls get compiled as regular jumps (like a tail function
calls). But sometimes the compiler will insert jumps in the middle of a
function and this will confuse things. To disambiguate, we have to check
that we are jumping to the start of a function.