Fix perf regression introduced by the addition of page events
Doing a `.clear()` on a vector retains the allocated memory, doing a
`mem::take` will consume it and cause an extra free and malloc when we
use the collection again.

Instead of doing `mem::take` we just inspect the allocated memory before
calling `.clear()`

This returns the execute benchmark on my Linux container to the speed as
it was before the original change (14Mhz -> 17Mhz) and hopefully
accounts for all the loss in the recent official benchmark runs.