Store regions as aligned words instead of bytes
Currently, when storing to a byte region in the guest, for each byte to
be stored, it reads the word which overlaps with its address, then
stores that word back with the single byte replaced. Because these
regions are always word-aligned, in practice, this peeking logic can be
removed and the words can be directly written.

This allows the memory in the guest region to be uninitialized.