zkvm: enable 3gb of guest memory
This is a simple fix to adjust the GUEST_MAX_MEM. There maybe more that
we can do to clean up risc0-zkvm-platform but this could result in
semver breaking changes such as removing public symbols. This might be
more involved so this change is meant to do the bare minimum to enable a
larger heap for the guest.