ZKVM-1125: zkvm: enable 3GB of guest memory (#2866)
This is a simple fix to adjust the GUEST_MAX_MEM. There maybe more that
we can do to clean up risc0-zkvm-platform but this could result in
semver breaking changes such as removing public symbols. This change is
meant to do the bare minimum to enable a larger heap for the guest.

Resolves ZKVM-1107

---------

Co-authored-by: Frank Laub <flaub@risc0.com>