Backport 1.1: ZKVM-1148: TraceEvent handles unknown events from futur… (#2894)
…e versions gracefully (#2886)

If a newer version of the circuit adds a new event type, the older
version of zkvm should just ignore it if received. Since this is
expected behavior, the fact that these events are being ignored is only
being logged at trace-level.

This change will need to be backported to 1.3, 1.2, and 1.1