Make 1.1 branch crates depend on 1.1.X of risc0-zkp
We made a breaking change to risc0-zkp in 1.2.X.
The 1.1.X crates won't compile with that version of risc0-zkp.

Looking on crates.io, of published crates pretty much only risc0 crates
depend on risc0-zkp directly, so I think it is sufficient for now to
just patch this dependency to fix this issue for any lingering 1.1.X
users.

This also just helps us verify that other changes to this patch release
compile correctly with future versions.