Grammatical suggestions for: README.md
Hi,

I have two suggestions for this file:

First; There is a small issue with "complex workspace or directory structure." It would be better to use "a" to make the sentence more grammatically correct: "However, this crate fails to build when placed in a complex workspace or directory structure",

Second: The use of "panic" could be more consistent. You could use: "This example does not produce a receipt and is intended as a test to ensure that the guest code runs without panicking".

Thanks.