ZKVM-1195: Fix bug with rzup concerning r0vm version containing `-` (#2939)
This was a little bit of a convoluted bug.

When we were checking for the existence of `r0vm` being installed, we
were asking the path-finding code to look for directories with `r0vm` in
the path, this was a bug.

This bug was masked though by the fact that the extension finding part
had a fallback part where it would just parse the version from the
string until the first `-`. Since versions with `-rc` in them have this
`-` it broke this logic.

This fallback part never needed to exist, we could instead just pass the
right component into this function. This fixes the bug because the
regular version finding logic works correctly using the component name
as a way to find the version number.