Update `cargo risczero new` to use versions of crates it was compiled with
Since we are changing the way we version crates to not be the same for
everything in one release, we need a way for `cargo risczero new` to
select the correct version for the current release. One way to do this
is to have it just use the versions of the crates it was compiled with.

Once we are using the same version as the release version on main
running one compiled from main will work the same as the release, but
right now it uses the `-alpha.1` version of crates which does work
because the `.rc.1` crates satisfy that constraint somehow.