ZKVM-1208: Update crate version numbers with new scheme (#2957)
In the new crate version numbering scheme each crate will have its own
version.
The crates on main will either:
- have the version of the last release if there are no changes since
that release
- have the version of the next release if there are changes since the
last release

We are going to bump all the version numbers for the next release though
(but not all major version numbers), to make things easier to start but
also to avoid associating with the 1.3.0 release.

I am also working on changes to the CI semver checks so that these rules
are always respected and the PR with changes to a crate will include
bumping the version number. Before I can make those checks pass though,
I need the main version numbers to be right.