ZKVM-1222: Implement new SemVer checks (#2968)
This patch implements the new SemVer checks we have discussed as a new
`xtask` command

It first runs `cargo-semver-checks` via its Rust API normally, then it
adds our check for patch releases by doing a new check.
This new check basically runs `cargo publish` on crates where the
current version number matches the published version number and compares
the version that would be published with the version that is published
and if there are any changes, we require a bump of the version number.

This PR also enables the new checks in CI replacing the existing checks.
It doesn't move to make it be required yet, I thought we would try it
out for a bit before blocking commits behind it.