[release-2.0]: Add missing -rc.2 to 3 crates (#2983)
I made a mistake and somehow got the version number wrong for these
crates. I want to fix it in crates.io by republishing these crates with
the right version and yanking the old version. This is in large part due
to fixing the semver checks, but also for posterity.