ZKVM-1229: Add unchecked/checked versions of ext field ops (#2977)
Split the extension field precompile operations into checked and
unchecked versions that do or don't (respectively) check whether any
component of the result is >= the base field prime.

The unchecked version is still constrained to give a result equivalent
to the checked result modulo the prime modulus; the check is just to
ensure that the result is a representation of the field element
"normalized" to its unique value `[0, p)`.

---------

Co-authored-by: iddo <iddo@risczero.com>