Fix an issue with sys_read where it would read past the end of the supplied buffer, or fail in other ways.
The branch that had the issue was in the v2 kernel and only taken when
the read is greater than 1024. By beefing up the test to do larger reads
we can see the bug trigger.

The sys_read syscall would pass arguments in registers a0-a4. The first
3 registers (a0, a1, a2) are common to all the ecall_read syscalls.
These are `syscall_ptr`, `buf`, and 'len`. The last two arguments (a3,
ar) are `fd` and `nwords`.

On v2, we implement ecall_read differently where it accepts `fd`, `buf`,
`len` as the first three common arguments. We end up adapting the API in
the v2 kernel. The kernel calls the host ecall_read from kernel code. We
stash the `syscall_ptr` in `fd` (the new ecall_read's `fd` argument not
to be confused with sys_read's `fd` argument) and `buf` and `len` in
their respective arguments, these are passed via kernel registers.

In the host ecall_read handler we unpack and adapt these arguments (the
kernel registers a1 and a2) before calling into the syscall_read code
(which we know to do thanks to a0) passing the two needed values (`buf`
and `len` directly through).

Once in the syscall code we read the two other arguments, the actual
`fd` and `nbytes` from the user register directly.

The return values are also different and there is code to adapt that
including another ecall_read call to ferry the result, but its not
relevant to the issue.

This all works fine, until we deal with reads greater than 1024. The new
ecall API has the limitation of only doing reads <= 1024. So if we do
reads >= 1024 the v2 kernel has to break up the read into smaller reads.
This happens in the Rust implemented part of the kernel, and this is
only taken for sys_read, the other syscalls aren't able to deal with
chunking at all.

When we do this chunking though, we forget about the other sys_read
arguments, in particular the value stored in a4. This value also
contains the length and needs to be updated. By not updating it, the
sys_read call does unexpected stuff, including reading not enough or too
much.