ZKVM-1359: Remove extra copies from extract_zkr (#3123)
There are extra copies going on here when we realloc the `Vec` inside
`read_to_end` and when we call `Vec::from`.

The larger the zkr the more of an issue this is. The zip format should
have the size of the entry in it, so we can do better.

Additionally the cast from `[u8]` -> `[u32]` is one of higher alignment
and kinda questionable. bytemuck will crash if the alignment is not high
enough and likely or allocation aligns all allocations at > 4, but if we
use a different memory allocator in the future this might not be true.

This was just something small I noticed when reading the recent
recursion witgen GPU patch.