ZKVM-1379: Add gdb support (#3164)
This adds a way to run the zkvm executor with a gdb server. This allows
users to attach gdb to the executor and set breakpoints and do
backtraces. This is enabled by passing `--with-debugger` to `r0vm` or
calling `run_with_debugger` on `ExecutorImpl`.

The riscv32im gdb must be used and I added code to rzup to distribute
it.

In order to create a portable version of gdb which works on multiple
releases of Ubuntu (and likely other distros) I had to disable its
Python support. This is a bit of a shame since I think the official Rust
gdb plugins are written in Python.

I compiled gdb using our existing riscv gnu tools repository. I added a
PR which updates the README to include instructions for building
yourself. https://github.com/risc0/toolchain/pull/3
I added the compiled binaries to the current release for the toolchain
repo and this is where rzup downloads it from.

This is just an initial version, there are still many more features you
might imagine it having.