ZKVM-1405: [actor-system] Pipeline tasks in the worker  (#3194)
This upgrades the actor system TCP communication to have multiplexing,
then it uses this new feature to pipeline tasks in the worker. It grabs
the next one while it is working on the current one.

I will need to merge https://github.com/risc0/risc0/pull/3191 with this
one. I'm hoping to merge it on top of this probably because I think it
will be easier that way, since I have to probably redo the tcp metric
stuff.