rzup: Implement installation of risc0-groth16 component (#3268)
This implements downloading and installing the new `risc0-groth16`
component.

### The `risc0-groth16` component

This new component uses S3 to store its metadata and tar.xz files. This
is the first component to do this, and a new S3 distribution has been
added to support this. If this works out well we can transition existing
components to S3 and not have to deal with downloading from GitHub
anymore.

The risc0-groth16 component is implemented as "target agnostic" meaning
the same bits are used on all platforms. The code for a target specific
S3 component has been added but nothing is using it yet.

When this component is selected as default, no symlinks are created,
only the `rzup` metadata is changed. Unclear that this default version
for this component will be used for anything, if not it is fine.

I uploaded to S3 a dummy tar.xz for now under the version 1.0.0-beta1.
We can edit the bucket to add the real files.

### S3 backed `rzup` components

The S3 distribution stores a `distribution_manifest.json` per component
that contains a list of releases keyed by version and shasums for the
artifacts. The artifacts are intended to be stored as objects at a
"sha256" prefix keyed off of their base64-encoded sha256 sum. This
should allow for releases to be obtained by the client atomically. Also
it should allow for verification of the artifacts contents later in the
client by verifying the sha256 sum.

I wish to add in a later PR a new `publish` subcommand to `rzup` which
will upload artifacts to S3 and update the manifest accordingly. We can
hopefully use this in our release GitHub action to store future releases
on S3 automatically.