Don't check RISC0_DEV_MODE env-var in guest (#3270)
Checking this environment variable was a needless mistake in the guest
because we ignore the result anyway. This caused guests doing
verification to require the sys-getenv feature, which is not what we
want.

In addition to fixing the check I improve the test coverage of verifying
receipts in the guest:
- Run our verify_guest example as a test (nothing seemed to be running
it before)
- Remove `sys-getenv` feature from guest program (not requiring this is
a goal)
- Add dev-mode feature/test for verify_guest (allowing this to work is
another goal)