actor system: Implement allocator (#3425)
This changes how the actor system does scheduling of work by
introduction of something new called the "allocator"

Both manager and workers connect and register with the allocator. API
requests go the allocator and it routes them to the appropriate manager.
Managers talk to the allocator to choose a worker to send work to.

Workers talk to the allocator to allocate hardware, both GPUs and CPU
cores.

The manager now sends work to workers rather than workers asking for
work from managers.

Workers connect to managers by way of communication with the allocator
so that now only the allocator needs to have well-known ports.

There is a new CLI tool for the actor system scheduling invoked via
`r0vm --actor-status <allocator-address>`. It provides a real-time
snapshot into what jobs workers are currently doing and will be doing
soon.